requirements.txt
PyQt6>=6.0,<7.0
biopython>=1.79,<2.0
